chore: Move test `design` to its own folder
