+++ crates/moon/tests/test_cases/design/lib/stack/lib.mbt
