+++ crates/moon/tests/test_cases/design/lib/stack/moon.pkg.json
