+++ crates/moon/tests/test_cases/design/lib/vec/moon.pkg.json
