+++ crates/moon/tests/test_cases/design/main1/main.mbt
