+++ crates/moon/tests/test_cases/design/main1/moon.pkg.json
