+++ crates/moon/tests/test_cases/design/main2/main.mbt
