+++ crates/moon/tests/test_cases/design/main2/moon.pkg.json
